K-JH closes #6 Update vagrant inventory
add groups;
  - etcd
  - master
  - minions
Separate the ip's of master and minions by at least 10 addresses.
When configuring the etcd network it is necessary to specify the
CIDR address, and although we can specify x.x.x.x/31. It is easier
to visualize if the master is x.x.x.d0 and the minions are
x.x.x.(d+1)x